% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECT.R
\name{ECT}
\alias{ECT}
\title{ECT}
\usage{
ECT(
  num_items = 18L,
  with_welcome = TRUE,
  take_training = FALSE,
  with_finish = TRUE,
  label = "ECT",
  feedback = ECT_feedback_with_score(),
  dict = ECT::ECT_dict
)
}
\arguments{
\item{num_items}{(Integer scalar) Number of items in the test.}

\item{with_welcome}{(Scalar boolean) Indicates, if a welcome page shall be displayed. Defaults to TRUE}

\item{take_training}{(Logical scalar) Whether to include the training phase. Defaults to FALSE}

\item{with_finish}{(Scalar boolean) Indicates, if a finish (not final!) page shall be displayed. Defaults to TRUE}

\item{label}{(Character scalar) Label to give the ECT results in the output file.}

\item{feedback}{(Function) Defines the feedback to give the participant
at the end of the test.}

\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
This function defines a ECT  module for incorporation into a
psychTestR timeline.
Use this function if you want to include the ECT in a
battery of other tests, or if you want to add custom psychTestR
pages to your test timeline.
}
\details{
For demoing the ECT, consider using \code{\link{ECT_demo}()}.
For a standalone implementation of the ECT,
consider using \code{\link{ECT_standalone}()}.
}
