% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{ECT_feedback_with_score}
\alias{ECT_feedback_with_score}
\title{ECT feedback (with score)}
\usage{
ECT_feedback_with_score(dict = ECT::ECT_dict)
}
\arguments{
\item{dict}{The psychTestR dictionary used for internationalisation.}
}
\description{
Here the participant is given textual feedback at the end of the test.
}
\examples{
\dontrun{
ECT_demo(feedback = ECT_feedback_with_score())}
}
